Add PickerGifView with basic gif showing functionality.
This is for showing gifs in the new ChromeOS picker feature. Various
details are still TBD (e.g. how many gifs can be shown, whether to
always animate or only animate on hover, resolution requirements), so
just add some basic functionality to experiment with for now.

Gifs are planned to be fetched from Google's tenor API:
https://developers.google.com/tenor

Then, the gifs are decoded out of process using the data_decoder
service, via ash::image_util API. The decoded frames are stored in
PickerGifView, which handles frame updates. Note that the gifs are
downscaled when they are decoded if needed to reduce the total size of
all frames to `kMaxImageSizeInBytes`:
https://source.chromium.org/chromium/chromium/src/+/main:ash/public/cpp/image_util.cc;l=156;drc=4597c83760e5b831ec153cba625e571d0a4ec6f9

Bug: b:316817118, b:316936723

Change-Id: Ia1bed51d6a640bb4aa1dbfc5ee1d823b35fa9564
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5124918
Reviewed-by: Scott Violet <sky@chromium.org>
Reviewed-by: Theodore Olsauskas-Warren <sauski@google.com>
Commit-Queue: Michelle Chen <michellegc@google.com>
Reviewed-by: Darren Shen <shend@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1243889}