[FspCache] Introduce CachedFileSystem
This is currently just a pass through wrapper around ProvidedFileSystem
that simply adds additional logging capabilities. It is only enabled
when the feature flag is on and will serve as the basis for intercepting
calls to the FSP.

VLOG(1) have been added for operations which are likely to affect a
local content cache, whilst VLOG(2) exist for any operation which (even
when a cache is implemented) will likely pass through to the underlying
ProvidedFileSystem.

Bug: b/317137739
Change-Id: I6f06b3b8e8b5165bc63237e6640a52c23985b146
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5172831
Reviewed-by: Cassy Chun-Crogan <cassycc@google.com>
Commit-Queue: Ben Reich <benreich@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1243892}