[ssci] Update fdlibm metadata
This CL updates the URL for fdlibm, so that it is an absolute URL and
actually links in chrome://credits instead of being a relative link.

Also updated Version to N/A, as the dependency is versioned by Revision
instead.

Bug: b:290296624
Change-Id: I0a9a95efb27fbdf49e4122352cc7ef4fb83e1624
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5162127
Commit-Queue: Joey Scarr <jsca@google.com>
Reviewed-by: Joey Scarr <jsca@google.com>
Auto-Submit: Anne Redulla <aredulla@google.com>
Reviewed-by: Michael Wilson <mjwilson@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1243906}