Fix ProfileTokenQuality::AddObservation() CHECK() failure
In AddObservation(type) the CHECK() that type is among the supported
types of *profile_ is failing (even after crrev.com/c/5169107 - see
linked crash report in the bug). This seemed strange, since the very
same check is performed earlier in the call stack in
AddObservationsForFilledForm().

Turns out the bug is in AddSubsetOfObservations(observatiosn), which is
supposed to only add a random subset of observations for privacy
reasons. The function works by randomly shuffeling the observations and
calling AddObservation() for the first k, where k is the result of a
function agreed upon with privacy. While computing k, the code
unfortunately references not only observations, but also observations_,
the member variable storing the already existing observations. As a
result, k can be larger than observations, causing code below to reach
out of memory, adding garbage observations of not necessarily
supported types.

This is actually a test for this case, but it had a bug too: The test
created a form with 12 different types, filled it with a complete
profile and submitted. It expected that only k = 8 (instead of the usual
k = 12 - 3 = 9 for "regular" sized forms) observations get added.
Turns out that one of the fields got dropped by some filtering logic
earlier, since it wasn't filled, because the "complete" profile doesn't
have two last names. Instead, the now uses NAME_MIDDLE and NAME_LAST,
which the profile has.

Fixed: 1515409
Change-Id: Ibb4fac62c7b5885f13827d629b46c3b2af800027
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5173598
Commit-Queue: Florian Leimgruber <fleimgruber@google.com>
Commit-Queue: Jihad Hanna <jihadghanna@google.com>
Auto-Submit: Florian Leimgruber <fleimgruber@google.com>
Reviewed-by: Jihad Hanna <jihadghanna@google.com>
Cr-Commit-Position: refs/heads/main@{#1243968}