Delay navigation until app foreground after external app launches
When an external application is launched chrome is not always put
in background (this is notably the case with system application that
show a confirmation popup to the user).
In that case, a quick navigation when Chrome is inactive foreground
could lead to spoofing.

This CL postpone navigations until the Chrome is back in active state
to avoid showing the wrong page to the user.

Bug: 942807
Change-Id: I11635ba6827b9572a99058c0781aa0f978cc3a34
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5158433
Reviewed-by: Rohit Rao <rohitrao@chromium.org>
Reviewed-by: Quentin Pubert <qpubert@google.com>
Commit-Queue: Olivier Robin <olivierrobin@chromium.org>
Reviewed-by: Ali Juma <ajuma@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1244012}