Implement the non-deprecated builder.
Any class inheriting `BrowserContextKeyedServiceFactory` should
implement either `BuildServiceInstanceForBrowserContext` or
`BuildServiceInstanceFor`. The second one was deprecated,
this CL replaces it with the first one.

In order to be usable with std::make_unique, constructor and
destructor had to became public. This is consistent with most other
keyed services.

Bug: 1396138
Change-Id: I3ee1fa17c5005176c18eb5f57b8d21fcfd366591
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/4801344
Reviewed-by: Nathan Parker <nparker@chromium.org>
Commit-Queue: Arthur Milchior <arthurmilchior@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1244019}