[ios-blink] Minimal AudioInputStream low latency implementation
A minimal implementation of AudioInputStream has been added to this CL
to make recording work. In the current minimal implementation, only
mono channel input streams are supported.

This implementation was manually tested on iPhone 13 with iOS 17.1.1

This change builds on the work done in below earlier CLs.
- https://crrev.com/c/5098306
- https://crrev.com/c/5103948
- https://crrev.com/c/5106188
- https://crrev.com/c/5109204

TEST: ios-deploy --justlaunch --bundle \
      out/Debug-iphoneos/content_shell.app --args=" \
      --auto-accept-camera-and-microphone-capture \
      https://web-recorder.glitch.me"

Change-Id: If4bee7e841fc0a734a2a7ffca5d5a69e1a3a83d0
Bug: 1413450
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5007480
Commit-Queue: Abhijeet Kandalkar <abhijeet@igalia.com>
Reviewed-by: Thomas Guilbert <tguilbert@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1244049}