Move content/browser/renderer_host/input/input_router_impl.h into //content/common/input
This refactoring is a part of bigger project to handle input on viz
thread (go/input-on-viz). We are moving input related interfaces that
RenderWidgetHostImpl uses to //content/common/input. This would provide
a single stack in //content/common/input that we quickly enter in either
the browser or GPU viz which would allow for relatively clean movement
of code. GPU process can link in the same input stack and forward the
data to the renderer using the same code.

Bug: b/305903095
Change-Id: I59085c5f5e3fd7159cbab481e90ff8dba2f33fe9
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5159007
Reviewed-by: Alex Moshchuk <alexmos@chromium.org>
Reviewed-by: Eric Seckler <eseckler@chromium.org>
Commit-Queue: Aman Verma <amanvr@google.com>
Cr-Commit-Position: refs/heads/main@{#1244061}