[iOS] Promo to upload bookmarks should be limited to left-behind ones
The batch upload promo from bookmarks manager should not be shown if the
user has already uploaded the sync-left behind bookmarks.
This should be the solution to avoid the confusion that happens because
both local and sync left-behind bookmarks are in the local store.

To guarantee that the promo should never be shown after the user
chooses to upload these bookmarks, this CL introduces a pref for that,
and writes/reads to it.

Note: the left-behind bookmarks in this usage are: all the bookmarks
that exist in the local store, where the current account is the last syncing account, till the user uploads them using this promo.

Fixed: 1513110
Change-Id: Id5fef0ed0d0a88f42401e3c7ad7a3699dcff63c7
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5164234
Reviewed-by: Jérôme Lebel <jlebel@chromium.org>
Reviewed-by: David Roger <droger@chromium.org>
Commit-Queue: Jood Hajeer <jood@google.com>
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Cr-Commit-Position: refs/heads/main@{#1244122}