[Android][Mfill] Hide webview popups on focus change
This CL reintroduces hiding the popup as removed in r1182073.

On WebView, this means that the popup is closed if a field is unfocused.
On Chrome, this changes nothing since the unfocus event already closes
the accessory — the new behavior only resets the data. This is correct.

Bug: 1515887
Fixed: 1509300
Change-Id: I65acd7e52a44a073e84ae55852c68841b29c7135
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5168885
Reviewed-by: Anna Tsvirchkova <atsvirchkova@google.com>
Reviewed-by: Jan Keitel <jkeitel@google.com>
Commit-Queue: Friedrich Horschig <fhorschig@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1244132}