Address ubsan failures in MultipartUploadRequest
LUCI bisection added a comment on https://crrev.com/c/5124759 that
we're seeing failures on the ubsan builder due to this CL. The fix
looks straightforward enough, so

The failures indicate that we're binding a reference to nullptr. This
is explicitly called out as possible in simple_url_loader.h, so this
CL replaces a null response body with an empty string.

Bug: 1093129
Change-Id: I82f43bd187a170e94e7d93dc8e9414e94deaa5eb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5171129
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Commit-Queue: Daniel Rubery <drubery@chromium.org>
Reviewed-by: Lily Chen <chlily@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1244150}