Add devtools.timeline category into etw keyword list
FireIdleCallback event used to be fired when etw session is enabled.
However devtools.timeline category was never in the etw keyword list.
It worked because of CategoryToKeyword function had a bug.
Now that bug has been fixed, devtools.timeline is needed in order to
get the FireIdleCallback event to be fired when etw session is on.

Bug: 1515718
Change-Id: Ie8e123096df64585dbf674bbaa4f032d58afc45f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5172365
Commit-Queue: Myungsub Kim <mskim@microsoft.com>
Reviewed-by: Etienne Pierre-Doray <etiennep@chromium.org>
Reviewed-by: Eric Seckler <eseckler@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1244176}