🧇 Make SearchEngineChoiceDialogService not get created on Android
This should not change anything as no code path should require it, but
this change officializes that the factory would always return nullptr
and makes some other checks from the component also indicate that the
platform is not supported.

We here also update some test to check whether the choice screen
or the new settings should be enabled with more granularity, ahead of
enabling the feature by default.

Bug: b:315126434
Change-Id: Ic76016aec9df63bdefedbd67c9baf526f6cafe54
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5163613
Commit-Queue: Nicolas Dossou-Gbété <dgn@chromium.org>
Reviewed-by: Boris Sazonov <bsazonov@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1244178}