[Downloads] Elide beginning of display URL on chrome://downloads
To avoid URL spoofs by truncation of very long download URLs, this CL
changes the elision of the display URL on chrome://downloads from
eliding the end to eliding the beginning. Because the eTLD+1 aka
registrable domain of the URL is at the end, we want to always display
that as it is security-relevant.

Note that eliding the beginning may truncate the scheme (http/https) of
the URL. This is acceptable because, due to insecure download warnings,
the user has already been alerted of any insecure (http) downloads, so
the scheme of the download URL is a less important signal.

Screenshots:
https://drive.google.com/drive/folders/1G5ChkHbM82HjFrHPVerzwf4UB1MmQmiZ

Bug: 1514925
Change-Id: I61639cdf4b6682a2ba8961822f22bd7621e8ad13
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5170825
Reviewed-by: Xinghui Lu <xinghuilu@chromium.org>
Reviewed-by: John Lee <johntlee@chromium.org>
Commit-Queue: Lily Chen <chlily@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1244238}