customization: Reorder event rewriters
The event rewriters should be in the following order:
1. Peripheral Customization
2. Pre-Rewritten Event forwarder
3. a11y rewriter

This order is necessary so a11y features can use peripheral
customization rewritten events. Additionally, the pre-rewritten event
forwarder must come after the peripheral customization rewriter so that
mouse button events are included in the event forwarder. The intention
of the pre-rewritten event forwarder is intended to bypass keyboard
rewrites which does not happen in peripheral customization.

Bug: b/241965717
Change-Id: I8701e104fd883c5247e990a53625affc851bf034
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5178526
Commit-Queue: David Padlipsky <dpad@google.com>
Reviewed-by: Jimmy Gong <jimmyxgong@chromium.org>
Reviewed-by: James Cook <jamescook@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1244261}