[Hub] HubProvider select tab model
Observe pane changes and set the correct tab model as active
accordingly. This could potentially live elsewhere, but this seemed
convenient because it can register the observer when the hub is
built and remain active along with the life of the Hub.

Bug: 1505772
Change-Id: I47fbb4b15d2e19382404930c6582f347a06d2b61
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5177150
Code-Coverage: findit-for-me@appspot.gserviceaccount.com <findit-for-me@appspot.gserviceaccount.com>
Commit-Queue: Calder Kitagawa <ckitagawa@chromium.org>
Reviewed-by: Sky Malice <skym@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1244307}