[OOPIF PDF] Fix claiming stream when it doesn't exist
Fix a crash that occurs when attempting to claim a PDF stream that
doesn't exist.

This crash occurs when a page contains an embed to a PDF, and the source
of the embed is changed to another URL. This triggers
ReadyToCommitNavigation(), which will attempt to claim a non-existing
stream. The code will incorrectly try to claim a stream if there's an
existing claimed stream.

This change makes the surrounding comments about claiming streams more
accurate.

Bug: 1445746, 1510361
Change-Id: If4a6216ba5a46a0c304c00466ef244136744fa3b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5177316
Commit-Queue: Andy Phan <andyphan@chromium.org>
Reviewed-by: Lei Zhang <thestig@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1244319}