[headless] Implemented code to center a headless window on Windows
HWNDMessageHandler::CenterWindow() is used to position and size
popup windows like bubbles, so it needs to be implemented in
headless mode for them to get the correct size.

The added test is also executed on Mac and Linux to ensure
bubbles are sized correctly there.

Drive by: removed NOTIMPLEMENTED_LOG_ONCE() in HWNDMessageHandlerHeadless::SizeConstraintsChanged()
replacing it with an appropriate comment.

Bug: 1505857, 1504346
Change-Id: I1912b8493d548a59351e33d48513239c9ed474f4
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5068521
Reviewed-by: Scott Violet <sky@chromium.org>
Reviewed-by: Andrey Kosyakov <caseq@chromium.org>
Commit-Queue: Peter Kvitek <kvitekp@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1244343}