[dPWA] Removing all required fields from web app protos
Required fields are widely considered harmful in protos [1]. This patch
removes them for the web app system, so Chrome will no longer crash on
reading corrupted data here and can instead ignore the offending apps.

This patch also adds checks in the parsing logic to ensure that any
missing fields are correctly handled.

One nuance here that is already used are the management type bools.
These default to `false` if they are not set, and that is already used
by the code as `false`. So these are not checked.

[1] https://protobuf.dev/programming-guides/dos-donts/#add-required

Change-Id: Icd49c16a42f9076a9f56dea5b3849b4187152a3b
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5166156
Commit-Queue: Daniel Murphy <dmurph@chromium.org>
Reviewed-by: Alan Cutter <alancutter@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1244402}