Refactor layout for launcher toast labels
The toast contains multiline labels, so its preferred size depends
on the width of the container - that means that if the toast parent
bounds change during layout, so could the preferred size for the toast.

To prevent preferred size changes mid layout, update the flow so the
app list bubble view propagates the available bounds to the toast
containers before layout pass.
Additionally, refactor how toast view lays out the label container.
The box layout does not do a good job laying out multi line labels,
which generally get cut-off. Part of the reason seems to be that the
label preferred size depends on the current label bounds, which may
later change during layout.
To work around the issue, this updates AppListToastView to calculate
it's preferred size without depending on the label preferred size -
depending on the available bounds, it will calculate the width that's
available to labels, and then use that to determine minimum height
needed to accommodate the label text.
Then in layout, the labels will be correctly sized so they fit their
text, before starting the toasts box layout logic (which will correctly
position labels along with other toast views).

BUG=b:317067114

Change-Id: I87f36ba324524517d43ced446cf57300c3d9523c
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5171132
Reviewed-by: Matthew Mourgos <mmourgos@chromium.org>
Commit-Queue: Toni Barzic <tbarzic@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1244414}