Image Loader: Create thumbnail with the video frame at the midpoint.
Currently the video thumbnail is created with the frame of the third
second. However, if the duration of the video is less than 3 seconds,
the thumbnail may be a empty frame. This CL uses the video frame at the
midpoint to create the thumbnail.

Bug: b:269717515
Test: Test 100 times on the videos whose duration are less than 3 secs.
Change-Id: Id0844d9803457bd8532afd5d429cc073c619ed29
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5175010
Reviewed-by: Peter Marshall <petermarshall@chromium.org>
Reviewed-by: Luciano Pacheco <lucmult@chromium.org>
Reviewed-by: Marcello Salomao <msalomao@google.com>
Commit-Queue: Sean Li <seannli@google.com>
Cr-Commit-Position: refs/heads/main@{#1244428}