cros: Speculative multitask menu nudge lacros crash fix
On lacros, we can try to display the nudge even if it is already shown.
We suspect this can happen because prefs are read and written to async
on lacros. For example:
  - User does action which would trigger nudge. Lacros reads prefs from
    ash.
  - User repeats action before the pref is read (or before new values
    are written to).
  - Pref is done reading and callback is sent for action 1. Nudge is
    shown.
  - Pref is done reading and callback is sent for action 2.

This doesn't happen on ash because read/write/animate is done in sync,
and the pref condition of showing once every 24hrs is more than
enough for the animation to complete.

I thought of adding some scheduling or weak pointer checks in the
lacros delegate, so we only ever query the pref service one at a
time, but it was much more complex and this edge case theoretically
should be very rare as the nudge only shows 3 times, and the user has
to activate the same window before the getting and setting the pref
is complete.

Test: manually using post task
Test: lacros_chrome_browsertests MultitaskMenuNudgeDelegateLacrosBrowserTest.QuickActivations
Bug: 1505602
Change-Id: I89b71ee17d4e60223f9feba8afc7c14cd335a7e5
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5084856
Reviewed-by: Hidehiko Abe <hidehiko@chromium.org>
Commit-Queue: Sammie Quon <sammiequon@chromium.org>
Reviewed-by: Ahmed Fakhry <afakhry@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1244430}