Prerender: stop prerender from overcounting not triggered failure
Before this CL, PrerenderHostRegistry records
BatterySaver/PreloadingDisabled etc every time omnibox receives a new
input, but this recording logic is not consistent to other final
statuses. The inconsistent logic results in extremely high failure
rate in UMA, but that does not really reflect the things we'd like to
know.
This CL rewrites this logic. WebContentsImpl::StartPrerendering will
always creates a new PrerenderHandle instance. For PrerenderHandles
whose frame_tree_node_id_ is invalid, it means that the prerender
attempt was rejected by pre-checks. But we keeps this instance alive
to track the historical attempts.


See this document for rational.
https://docs.google.com/document/d/1a1mlf_JxMD2018bRZtzCR0J0eqFPY_0myBOGdcV12ts/edit#heading=h.7omvz2uyeznq

Bug: 1505293
Change-Id: I55c6a1d521fdb7fd522a1ce808224cd5720abff1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5069215
Reviewed-by: Taiyo Mizuhashi <taiyo@chromium.org>
Reviewed-by: Alan Cutter <alancutter@chromium.org>
Commit-Queue: Lingqi Chi <lingqi@chromium.org>
Reviewed-by: Huanpo Lin <robertlin@chromium.org>
Reviewed-by: Rakina Zata Amni <rakina@chromium.org>
Reviewed-by: Hiroki Nakagawa <nhiroki@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1244460}