Add unused preload request count UMAs
This CL adds following UMAs:
- Renderer.Preload.UnusedResource2{Trigger}
- Renderer.Preload.UnusedResourceCount

Renderer.Preload.UnusedResource2 is the new version of
Renderer.Preload.UnusedResource. The older UMA sends ResourceType enum
via `UMA_HISTOGRAM_COUNTS_100`, which is not intuitive because the
recorded one is an enum value, not a counted number.. Also in the xml,
`enum=BooleanSuccess` is set and that is not appropriate usage.

This CL adds the total count to unused preload request as well.

Bug: 1498777
Change-Id: I7169571eb59ee135c9854523a04fc0965fc9220e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/5144086
Reviewed-by: Patrick Meenan <pmeenan@chromium.org>
Reviewed-by: Takashi Toyoshima <toyoshim@chromium.org>
Reviewed-by: Minoru Chikamune <chikamune@chromium.org>
Commit-Queue: Shunya Shishido <sisidovski@chromium.org>
Reviewed-by: Kouhei Ueno <kouhei@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1244496}